+++ caspy3/qt_assets/tabs/derivative.py
@@ -17,7 +17,7 @@
 #
 
 from PyQt5.QtCore import pyqtSlot, Qt
-from PyQt5.QtWidgets import QShortcut, QWidget
+from PyQt5.QtWidgets import QShortcut, QTextBrowser, QWidget
 from PyQt5.QtGui import QCursor, QKeySequence
 from PyQt5.uic import loadUi
 
@@ -166,18 +166,21 @@ class DerivativeTab(QWidget):
         super().__init__()
         self.main_window = main_window
         loadUi(self.main_window.get_resource_path("qt_assets/tabs/derivative.ui"), self)
+        self.eout: QTextBrowser = self.DerivOut
+        self.aout: QTextBrowser = self.DerivApprox
 
         # Shortcuts
         cshortcut = QShortcut(QKeySequence("Ctrl+Return"), self)
         cshortcut.activated.connect(self.calc_deriv)
         pshortcut = QShortcut(QKeySequence("Ctrl+Shift+Return"), self)
         pshortcut.activated.connect(self.prev_deriv)
-
         self.init_bindings()
 
     def init_bindings(self) -> None:
         self.DerivPrev.clicked.connect(self.prev_deriv)
         self.DerivCalc.clicked.connect(self.calc_deriv)
+        self.eout.mousePressEvent = lambda _: self.eout.selectAll()
+        self.aout.mousePressEvent = lambda _: self.aout.selectAll()
 
     def stop_thread(self) -> None:
         pass
@@ -197,10 +200,6 @@ class DerivativeTab(QWidget):
 
             self.DerivOut.setText(self.main_window.exact_ans)
             self.DerivApprox.setText(str(self.main_window.approx_ans))
-
-        # TODO: This
-        self.DerivOut.selectAll()
-        self.DerivOut.setFocus()
 
     def prev_deriv(self) -> None:
         self.DerivOut.viewport().setProperty("cursor", QCursor(Qt.WaitCursor))